70: Move out editable logic into separate component
